requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.84
botocore==1.31.84
charset-normalizer==3.3.2
distlib==0.3.7
Django==4.2.7
django-debug-toolbar==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
filelock==3.12.2
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
loguru==0.7.2
multidict==6.0.4
mysqlclient==2.2.0
platformdirs==3.10.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qasync==0.26.0
referencing==0.30.2
rpds-py==0.12.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.2
yarl==1.9.2
python-json-logger
